requirements.txt
streamlit
langchain-community
langchain-groq
sentence-transformers
chromadb
python-dotenv
PyPDF2
scikit-learn
numpy